get: return an empty object if host is not found
Previously, the credential helper spec called for an error if a token
could not be returned for `get`. However, Terraform makes
unauthenticated requests to registry.terraform.io. Previously, this
was just for modules, but now extends to all providers as well in
Terraform 0.13, making it even more visible:

https://github.com/hashicorp/terraform/issues/25938

The Terraform website will be updated to say:

> If the credentials helper definitively has no credentials for the
> given host, then it must print an empty JSON object to stdout and
> exit with status zero.

https://github.com/hashicorp/terraform/pull/26244

This implemenets that recommendation, returning an empty object when
keyring.Get returns a "key not found" error